+++ src/main/java/subway/domain/line/LineResponse.java
@@ -1,0 +1,41 @@
+package subway.domain.line;
+
+import subway.domain.station.StationResponse;
+
+import java.util.Collections;
+import java.util.List;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private List<StationResponse> stations;
+
+    public LineResponse() {
+    }
+
+    public LineResponse(Line line) {
+        this.id = line.getId();
+        this.name = line.getName();
+        this.color = line.getColor();
+
+        List<StationResponse> stationResponses = line.getAllStations().toResponse();
+        this.stations = Collections.unmodifiableList(stationResponses);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+}
