+++ src/main/java/subway/domain/section/SectionRequest.java
@@ -1,0 +1,34 @@
+package subway.domain.section;
+
+import subway.domain.station.Station;
+
+public class SectionRequest {
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
+
+    public SectionRequest() {
+    }
+
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Section toSection(Long lineId) {
+        return new Section(null, lineId, new Station(upStationId), new Station(downStationId), distance);
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
