+++ src/main/java/subway/domain/station/StationRequest.java
@@ -1,0 +1,16 @@
+package subway.domain.station;
+
+public class StationRequest {
+    private String name;
+
+    public StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
