--- src/main/java/subway/line/LineResponse.java
@@ -1,41 +1,0 @@
-package subway.line;
-
-import subway.station.StationResponse;
-
-import java.util.Collections;
-import java.util.List;
-
-public class LineResponse {
-    private Long id;
-    private String name;
-    private String color;
-    private List<StationResponse> stations;
-
-    public LineResponse() {
-    }
-
-    public LineResponse(Line line) {
-        this.id = line.getId();
-        this.name = line.getName();
-        this.color = line.getColor();
-
-        List<StationResponse> stationResponses = line.getAllStations().toResponse();
-        this.stations = Collections.unmodifiableList(stationResponses);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-}
