--- src/main/java/subway/station/Station.java
@@ -1,54 +1,0 @@
-package subway.station;
-
-import java.util.Objects;
-
-public class Station {
-    private Long id;
-    private String name;
-
-    public Station() {
-    }
-
-    public Station(Long id) {
-        this(id, null);
-    }
-
-    public Station(String name) {
-        this(null, name);
-    }
-
-    public Station(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Station station = (Station) o;
-        return Objects.equals(name, station.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-
-    @Override
-    public String toString() {
-        return "Station{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
-
